requirements.txt
# Web Framework
fastapi>=0.122.0
uvicorn[standard]>=0.38.0
python-multipart>=0.0.6
websockets>=12.0
sqlalchemy>=2.0.0

# Computer Vision
opencv-python>=4.12.0
Pillow>=12.0.0

# Object Detection & Deep Learning
ultralytics>=8.0.227
# For GPU (CUDA 12.1): Install with --index-url https://download.pytorch.org/whl/cu121
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Vision-Language Models & Fine-tuning
transformers==4.45.0
accelerate>=0.25.0
sentencepiece>=0.1.99
peft>=0.7.0  # Parameter-Efficient Fine-Tuning (LoRA)
bitsandbytes>=0.41.0  # For quantization

# Audio Processing
openai-whisper>=20231117
soundfile>=0.12.1
librosa>=0.10.1

# Vector Database & Embeddings
chromadb>=1.3.0
sentence-transformers>=5.0.0

# Task Queue & Caching
redis>=5.0.1
celery>=5.3.4

# Monitoring & Metrics
psutil>=5.9.0  # System metrics
py-cpuinfo>=9.0.0  # CPU info

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=2.0.0
pandas>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Socket.IO
python-socketio>=5.10.0

# Testing & Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0  # Coverage reporting
black>=23.0.0  # Code formatting
ruff>=0.1.0  # Fast linting
